requirements.txt
streamlit-authenticator>=0.4.2
streamlit
pandas
pyyaml
